Convert the database class to an object
I was going with the OOP approach since I thought it was simpler to test
and maybe it is for the unit test but not for the integration test.

When I rethink this, it might be a solution to rewrite the auto inject
directive to use different classes depending on if the system is under
test or not.